// src/constants/game/regular-wheel.enum.ts
export enum RegularWheelElement {
  ANY = "ANY",
  LEMON = "LEMON",
  PLUM = "PLUM",
  BELL = "BELL",
  CHERRY = "CHERRY",
  WATERMELON = "WATERMELON",
  KING = "KING",
  ORANGE = "ORANGE",
  PEAR = "PEAR",
  JOKER = "JOKER",
  GRAPE = "GRAPE",
  BAR = "BAR",
}

export type RegularWheel = RegularWheelElement[];